Organize contexts and name them properly - #7
