+++ app/javascript/components/blockouts/Blockout.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import moment from 'moment'
-import SchedulerContext from './scheduler-context'
+import SchedulerContext from 'components/blockouts/contexts/scheduler'
 
 class Blockout extends React.Component {
   renderTime = _ => {
