+++ app/javascript/components/blockouts/Calendar.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import SchedulerContext from './scheduler-context'
+import SchedulerContext from 'components/blockouts/contexts/scheduler'
 import PropTypes from 'prop-types'
 import DayPicker from 'react-day-picker'
 import 'react-day-picker/lib/style.css'
