+++ app/javascript/components/blockouts/EditBlockoutModal.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import SchedulerContext from './scheduler-context'
-import BlockoutFormContext from './blockout-form-context'
+import SchedulerContext from 'components/blockouts/contexts/scheduler'
+import BlockoutFormContext from 'components/blockouts/contexts/blockoutForm'
 import DateTimePicker from './DateTimePicker'
 import { RRule, RRuleSet, rrulestr } from 'rrule'
 import ReasonInput from './ReasonInput'
