+++ app/javascript/components/blockouts/ReasonInput.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import BlockoutFormContext from './blockout-form-context'
+import BlockoutFormContext from 'components/blockouts/contexts/blockoutForm'
 
 class ReasonInput extends React.Component {
   render () {
