+++ app/javascript/components/blockouts/Scheduler.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import DayPicker from 'react-day-picker'
 import moment from 'moment'
-import SchedulerContext from './scheduler-context'
+import SchedulerContext from 'components/blockouts/contexts/scheduler'
 import Calendar from './Calendar'
 import BlockoutList from './BlockoutList'
 import AddBlockoutButton from './AddBlockoutButton'
